Percentage-based selection of core determinants size.
This adds a toggle with which the user can use to set a fixed number of core
determinants (the old behaviour, defaulting to ncdets=100), or
alternatively a percentage-based number of core determinants. This
defaults to 90%, which captures the determinants of the wavefunction,
for a given iteration, which constitute 90% of the total wavefunction
weight. The goal of this mode is to better scale the number of core
determinants with the problem size.